requirements.txt
autopep8==1.6.0
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
firebase-admin==5.2.0
Flask==2.1.2
Flask-Cors==3.0.10
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-logging==3.2.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
proto-plus==1.22.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.14.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
six==1.16.0
structlog==21.5.0
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.1
